+++ src/Functional.CQS.AOP.Caching.Infrastructure/Invalidation/IReplaceFunctionalCacheItem.cs
