--- src/Functional.CQS.AOP.Caching.Invalidation/Functional.CQS.AOP.Caching.Invalidation.csproj
@@ -1,15 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFramework>netstandard2.0</TargetFramework>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <PackageReference Include="Functional.Primitives" Version="1.9.0" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\Functional.CQS\Functional.CQS.csproj" />
-  </ItemGroup>
-
-</Project>
