+++ src/Functional.CQS.AOP.IoC.PureDI.Caching/Invalidation/FunctionalCacheItemOverwriter.cs
@@ -3,6 +3,7 @@ using System.Collections.Concurrent;
 using System.Reflection;
 using Functional.CQS.AOP.Caching;
 using Functional.CQS.AOP.Caching.Infrastructure;
+using Functional.CQS.AOP.Caching.Infrastructure.Invalidation;
 using Functional.CQS.AOP.Caching.Invalidation;
 using Functional.CQS.AOP.IoC.PureDI.Caching.Models;
