+++ src/Functional.CQS.sln
@@ -45,11 +45,11 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Functional.CQS.AOP.IoC.Pure
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Functional.CQS.AOP.CommonTestInfrastructure.Caching", "Functional.CQS.AOP.CommonTestInfrastructure.Caching\Functional.CQS.AOP.CommonTestInfrastructure.Caching.csproj", "{D8D3B611-DC77-467D-A8CE-FB9225FCC2DF}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Functional.CQS.AOP.IoC.SimpleInjector.Caching", "Functional.CQS.AOP.IoC.SimpleInjector.Caching\Functional.CQS.AOP.IoC.SimpleInjector.Caching.csproj", "{CC3F1797-8BF2-4144-9254-79CD05D3FA5D}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Functional.CQS.AOP.IoC.SimpleInjector.Caching", "Functional.CQS.AOP.IoC.SimpleInjector.Caching\Functional.CQS.AOP.IoC.SimpleInjector.Caching.csproj", "{CC3F1797-8BF2-4144-9254-79CD05D3FA5D}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Functional.CQS.AOP.Caching.Invalidation", "Functional.CQS.AOP.Caching.Invalidation\Functional.CQS.AOP.Caching.Invalidation.csproj", "{A2B4C1D8-5270-4245-A41F-10872FC338A3}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Functional.CQS.AOP.IoC.SimpleInjector.Caching.Tests", "Functional.CQS.AOP.IoC.SimpleInjector.Caching.Tests\Functional.CQS.AOP.IoC.SimpleInjector.Caching.Tests.csproj", "{C3BA79AE-E134-45BD-85FD-3D27853284AF}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Functional.CQS.AOP.IoC.SimpleInjector.Caching.Tests", "Functional.CQS.AOP.IoC.SimpleInjector.Caching.Tests\Functional.CQS.AOP.IoC.SimpleInjector.Caching.Tests.csproj", "{C3BA79AE-E134-45BD-85FD-3D27853284AF}"
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Common", "Common", "{B236E576-7D86-4852-B44A-057FA63E0237}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -137,10 +137,6 @@ Global
 		{CC3F1797-8BF2-4144-9254-79CD05D3FA5D}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{CC3F1797-8BF2-4144-9254-79CD05D3FA5D}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{CC3F1797-8BF2-4144-9254-79CD05D3FA5D}.Release|Any CPU.Build.0 = Release|Any CPU
-		{A2B4C1D8-5270-4245-A41F-10872FC338A3}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{A2B4C1D8-5270-4245-A41F-10872FC338A3}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{A2B4C1D8-5270-4245-A41F-10872FC338A3}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{A2B4C1D8-5270-4245-A41F-10872FC338A3}.Release|Any CPU.Build.0 = Release|Any CPU
 		{C3BA79AE-E134-45BD-85FD-3D27853284AF}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{C3BA79AE-E134-45BD-85FD-3D27853284AF}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{C3BA79AE-E134-45BD-85FD-3D27853284AF}.Release|Any CPU.ActiveCfg = Release|Any CPU
@@ -152,12 +148,15 @@ Global
 	GlobalSection(NestedProjects) = preSolution
 		{487B1794-9128-4E56-976C-274D4268A5B4} = {3DB0C9FC-28E8-414D-8F16-B7C1EFF090E8}
 		{6A750B3C-029B-4932-955D-B2035DBAA757} = {3DB0C9FC-28E8-414D-8F16-B7C1EFF090E8}
+		{058D2CA3-35C1-4AAA-B459-269408BEC09B} = {B236E576-7D86-4852-B44A-057FA63E0237}
 		{4C7AEC53-6783-45E3-8E81-348521FCAE88} = {26B803A7-BF86-48AA-9668-F48626FF9075}
 		{34E2696C-9ED4-4CC1-959C-4779FCC67E1F} = {26B803A7-BF86-48AA-9668-F48626FF9075}
 		{04540474-8966-4E90-A90E-A7251F8B8118} = {3DB0C9FC-28E8-414D-8F16-B7C1EFF090E8}
 		{3B104AC4-23F1-4944-8B69-A13A812A3730} = {3DB0C9FC-28E8-414D-8F16-B7C1EFF090E8}
+		{8246DD38-6775-46D5-AF05-768647908E60} = {B236E576-7D86-4852-B44A-057FA63E0237}
 		{0A414F5B-E1EB-48A9-B58F-23380A6F151F} = {26B803A7-BF86-48AA-9668-F48626FF9075}
 		{EDB4D3EA-FDBA-4A23-A2BC-41541C352415} = {26B803A7-BF86-48AA-9668-F48626FF9075}
+		{D8D3B611-DC77-467D-A8CE-FB9225FCC2DF} = {B236E576-7D86-4852-B44A-057FA63E0237}
 		{CC3F1797-8BF2-4144-9254-79CD05D3FA5D} = {3DB0C9FC-28E8-414D-8F16-B7C1EFF090E8}
 		{C3BA79AE-E134-45BD-85FD-3D27853284AF} = {3DB0C9FC-28E8-414D-8F16-B7C1EFF090E8}
 	EndGlobalSection
